requirements.txt
playwright>=1.45.0
notion-client>=2.2.1
python-dotenv>=1.0.1
rich>=13.7.1
